Move LayerInfo to own file
